rajout de la supression de notif si tournoi annuler ou ready a tester

+++ user-service/src/menu/notifications/tournaments/userAnswerTournament.js
@@ -1,4 +1,4 @@
-import env from "../../../config/env.js";
+//import env from "../../../config/env.js";
 import { getSecret } from "../../../index.js";
 import prefix from "../../../tools/url.js";
 import tlsAgent from "../../../tools/tlsAgent.js";
@@ -9,48 +9,73 @@ export async function   userAnswerTournament(request, reply)
     const idUser = request.user.idUser;
     const { ownerSlug, tournamentName, answer } = request.body;
 
-    const ownerId = db.prepare("    SELECT \
-                                        id \
-                                    FROM \
-                                        users \
-                                    WHERE \
-                                        slug = ?").get(ownerSlug);
-
-    const tournamentId = db.prepare("   SELECT \
-                                            notif_tounament_id \
+    try
+    {   
+        const ownerId = db.prepare("    SELECT \
+                                            id \
                                         FROM \
-                                            notifications \
+                                            users \
                                         WHERE \
-                                            notif_tournament_name = ?").get(tournamentName);
-    db.prepare(     "DELETE FROM \
-                        notifications \
-                    WHERE \
-                        notif_user_id = ? \
-                    AND \
-                        notif_sender_id = ? \
-                    AND \
-                        notif_tournament_id = ? \
-                    AND \
-                        notif_type = 'tournament_invite'").run(idUser, ownerId, tournamentId);
-    let url;
-    if (answer === "decline")
-        url = `${prefix}://session-service:${env.session_port}/tournament/decline`
-    else
-        url = `${prefix}://session-service:${env.session_port}/tournament/accept`
-    answerTournament(userId, ownerId, tournamentId, tournamentName, url);
+                                            slug = ?").get(ownerSlug);
+        if (!ownerId || !ownerId.id)
+            return reply.code(404).send({ error: "Resource not found ICI " });
+
+        const tournamentId = db.prepare("   SELECT \
+                                                notif_tournament_id \
+                                            FROM \
+                                                notifications \
+                                            WHERE \
+                                                notif_tournament_name = ?").get(tournamentName); //PROETGER AUSSI POUR LE TOURNAMENT ID
+
+        if (!tournamentId || !tournamentId.notif_tournament_id)
+            return reply.code(404).send({ error: "Resource not found ID TOUANMENT" });
+
+        console.log('\ntournamentId : ', tournamentId);
+        console.log('\nownerId : ', ownerId);
+        console.log('\nidUser : ', idUser);
+        db.prepare(     "DELETE FROM \
+                            notifications \
+                        WHERE \
+                            notif_user_id = ? \
+                        AND \
+                          notif_sender_id = ? \
+                        AND \
+                            notif_tournament_id = ? \
+                        AND \
+                            notif_type = 'tournament_invite'").run(idUser, ownerId.id, tournamentId.notif_tournament_id);
+        let url;
+        if (answer === "decline")
+        {;
+            //url = `${prefix}://session-service:${env.session_port}/tournament/decline`;
+            url = `http://session-service:3004/tournament/decline`;
+        }
+        else
+        {
+            //url = `${prefix}://session-service:${env.session_port}/tournament/accept`;
+            url = `http://session-service:3004/tournament/accept`;
+        }
+        answerTournament(idUser, ownerId.id, tournamentId.notif_tournament_id, tournamentName, url);
+        return reply.code(200).send();
+    }
+    catch(err)
+    {
+        return reply.code(500).send({ error: "Internal Server Error" });
+    }
 }
 
-export async function answerTournament(userId, ownerId, tournamentId, tournamentName) 
+export async function answerTournament(userId, ownerId, tournamentId, tournamentName, url) 
 {
     const api_key = getSecret('api_key');
 
+
+    console.log('info avant le FECTH: ', userId, ownerId, tournamentId, tournamentName, url);
     const res = await fetch(url, 
     {
         method: 'POST',
         headers: 
         { 
-            'Content-Type': 'application/json',
-            'x-internal-api-key': api_key
+            'x-internal-api-key': api_key,
+            'Content-Type': 'application/json' 
         },
         body: JSON.stringify(
         {
